chore(appliance): extract envtest helper package
In preparation for reuse.
