+++ internal/appliance/reconciler/BUILD.bazel
@@ -105,22 +105,20 @@ go_test(
     tags = [TAG_INFRA_RELEASE],
     deps = [
         "//internal/appliance/config",
+        "//internal/appliance/k8senvtest",
         "//internal/appliance/yaml",
-        "@com_github_go_logr_stdr//:stdr",
         "@com_github_life4_genesis//slices",
+        "@com_github_sourcegraph_log//logtest",
+        "@com_github_sourcegraph_log_logr//:logr",
         "@com_github_stretchr_testify//require",
         "@com_github_stretchr_testify//suite",
-        "@io_bazel_rules_go//go/runfiles:go_default_library",
         "@io_k8s_api//core/v1:core",
         "@io_k8s_api//rbac/v1:rbac",
         "@io_k8s_apimachinery//pkg/apis/meta/v1:meta",
         "@io_k8s_apimachinery//pkg/runtime/schema",
         "@io_k8s_client_go//kubernetes",
-        "@io_k8s_client_go//kubernetes/scheme",
         "@io_k8s_sigs_controller_runtime//:controller-runtime",
         "@io_k8s_sigs_controller_runtime//pkg/client",
-        "@io_k8s_sigs_controller_runtime//pkg/envtest",
-        "@io_k8s_sigs_controller_runtime//pkg/metrics/server",
         "@io_k8s_sigs_yaml//:yaml",
     ],
 )
