+++ internal/appliance/reconciler/reconcile.go
@@ -2,6 +2,7 @@ package reconciler
 
 import (
 	"context"
+
 	appsv1 "k8s.io/api/apps/v1"
 	"k8s.io/apimachinery/pkg/types"
 	"sigs.k8s.io/controller-runtime/pkg/predicate"
